+++ CHANGELOG.md
@@ -19,7 +19,8 @@ January 23rd, 2022
 - Removed the existing authentication system and replaced it with (DropAuth)[https://v0lttech.com/dropauth.php]
 - Simplified stylesheets by creating a main stylesheet that all themes have in common.
 - Organized the configuration back-end.
-- Re-designed the post database back-end.
+- Re-designed the post database back-end to make future expansion much easier.
 - Updated the way posts are viewed.
     - A brief summary of each post is now displayed on the main page.
     - Posts can be opened on a separate page to read the full article.
+- Authorized authors can now be configured independently of the admin account.